Improvements to the ValueElement and Material API
This changelist makes a number of improvements to the ValueElement and Material API, providing access to commonly-needed data without requiring the developer to understand the full MaterialX rule set.

From a high-level, the following methods and classes have been added:

- The ValueElement methods getBoundValue and getDefaultValue have been added, allowing uniform values to be easily evaluated in the context of a particular material.
- The Material methods getPrimaryShader[NodeDef/Name/Parameters/Inputs] have been added, allowing clients to easily access properties of the first shader that matches their requirements.
- The StringResolver class has been added, allowing clients to control how string modifiers (e.g. file prefixes and geometry-based substitutions) are applied to data values.

Note that some methods have been renamed or removed in this changelist, so code updates may be required when upgrading to this build of MaterialX.  See CHANGELOG.md for the full list of API changes.