Add GLSL subsurface scattering approximation
This changelist adds a simple subsurface scattering approximation to MaterialX GLSL.  The high-level approach is to integrate Burley's diffusion profile over the sphere using Eric Penner's curvature-based approximation.  While this ignores the thickness of the scattering volume, it keeps the implementation within a single shading pass and provides a meaningful visual interpretation of the subsurface scattering inputs.