Fixes for Material baking and Viewer refactor. (#1064)
* Update drawScreenSpaceQuad to accept optional Attibute id
The drawScreenSpaceQuad assumes
i_position at 0 and i_texcoord_0 at 1
But on macOS/GL Core Profile the locations are different.
e.g: i_position 1 and i_texcoord_0 0
This change queries the program for the input location.

* GlslRender and Viewer Updates
The MaterialX Viewer uses NanoGui and was making a call into GlslRender. This was crossing the two renderer states.
Instead now we
- Make GlslRender drawScreenSpaceQuad a member function instead of static
  This can now use the current active _program
- Add a helper renderScreenSpaceQuad to MaterialX Viewer
  This can use use the current material shader to setup attributes for Albedo Table and Shadowmap
* Updates to remove unused NSWindow helper and MaterialX classes to render quad.
* Add QuadMesh Utility in GeometryHandler
Utility in GeometryHandler to create a simple quad mesh.
This enables us to replace with stock geometry in future.