Remove viewer dependency on NanoGUI::GLShader
This changelist removes the dependency of the MaterialX viewer on NanoGUI::GLShader, allowing future iterations of the viewer to take advantage of other rendering frameworks (including ImGui and the modernized NanoGUI used in Mitsuba 2).

In place of NanoGUI::GLShader, shader compilation, introspection, and binding is now entirely handled by MaterialX::GlslProgram, unifying the approach to GLSL shading across the viewer, texture baker, and render test framework.