ESSL code generator addition (#693)
Web merge part 2: ESSL code generator

Previous check-in has modified other GLSL code snippets and shader node code generators to produce ESSL and GLSL compliant code such as avoiding explicit casts. This formally introduces a new generator to which shares existing GLSL code but is injected in into the final shader in a custom manner.

This change includes the following:
1. Refactors some of the underlying code in the GLSL generator to be modularized into virtual methods which can be overridden by the derived ESSL shader class.
2. Some pertinent points including modifying the language directives, and not initializing any uniforms.
3. From a usage perspective:
    - Include Python wrappers for this generator and add in a Python utility to allow arbitrary language output. (`generateshader`)
    -  Add the ability to output ESSL source code from MaterialXView
4. ESSL implementation checking uses `generateshader` as part of CI.
5. `glslang` is setup as part of CI with `glslangValidator` being invoked for ESSL and GLSL syntax checking of vertex and pixel shader stages. (Note: This covers all of the `pbr` library nodes + some stdlib nodes based on usage in the examples materials. 
basic infrastructure is now available.)
(Note: Actual WASM testing is not part of this check-in but will use this generator. ) As part of this CI has been modified a bit to reduce the number of builds for pushes vs PRs.
6. Some minor additional fixes for GLSL / ESSL source code implementation compliance in the pbr library.