Fix skipped inputs in createUIPropertyGroups
This changelist fixes a bug that caused shader inputs to be skipped in createUIPropertyGroups.

Instead of computing breadth-first indices in a dedicated (and inaccurate) method, we now compute exact depth-first indices by traversing the document tree in advance.

The original inaccurate method has been removed, as it is otherwise unused in the MaterialX codebase.