Initial implementation of Worley noise for OSL (#810)
Adds an initial implementation of Worley noise for OSL.

The hash is not yet identical to the one used by the GLSL and MDL implementations (an unsigned int right shift operator is missing in OSL). As discussed we will ignore this for now and we will refine this later to get identical results.