Ensure libMaterialXRenderHw has a dependency on libXt (#753)
After linking a program against a MaterialX build I made recently on Linux, I got this error on execution:

Error: /blah/libMaterialXRenderHw.so: undefined symbol: topLevelShellWidgetClass

On inspection, the problem was that `libMaterialXRenderHw.so` wasn't declaring a dependency on `libXt.so`. When I inspected the CMake source, I found this was because RenderHw wasn't doing a `find_package(X11)` - so that even though it calls `target_link_libraries(${X11_Xt_LIB})`, it does nothing because `${X11_Xt_LIB}` hasn't been populated yet.