Clarify usage of alpha and roughness
This changelist clarifies the usage of the terms alpha and roughness in the PBR libraries for GLSL and OSL, using the term 'alpha' to mean the square of artist-facing roughness in the context of specular BSDFs.

In the future, we should consider modifying the interfaces of specular BSDF nodes such as <dielectric_bsdf> and <conductor_bsdf> to follow this convention as well, since their "roughness" inputs actually represent alpha.

Additional changes:
- Align handling of input alpha between GLSL and OSL.
- Remove unused helper functions in the PBR library for OSL.