Remove hardcoded search path string
This changelist removes a hardcoded inclusion of the "libraries" string in calls to GenContext::registerSourceCodeSearchPath, giving control back to include directives to specify which top-level folder is intended.

As a result, applications with custom code generators will need to remove this string from their own calls to GenContext::registerSourceCodeSearchPath when they upgrade to MaterialX 1.38.4.

Supporting changes include:
- Give precedence to user-supplied libraries over the standard libraries in MaterialXView, generateshader.py, and translateshader.py.
- Minor fixes to texture baking in MaterialXView.