Add alpha to transparency detection
This changelist adds "alpha" to the list of shader inputs that are checked in transparency detection functions (e.g. mx::isTransparentSurface), making them inclusive of the conventions of the glTF PBR shading model.  Materials with non-zero or spatially-varying values for "alpha" will now be considered transparent by renderers that leverage these functions.