Track File Dependencies in ShaderStage
This change allows shader generator code to look at file dependencies in ShaderStages.
This is done by adding the file path of a SourceCodeNode to the ShaderStage's set of dependencies whenever a function definition is emitted.
Also, ShaderStage is changed to allow returning the set of include files that it requires.
Includes a unit test to check that a material's shader dependencies actually exist as files.