Improved path generation for glTF geometry loader (#1028)
1. Add in proper path generation when loading in glTF files. For now no "/" root is ever added and mesh names are
not part of the path, unless the parent transform path is empty. Previously incorrect paths could be produced, especially
as transform names were "generated". 

2. As part of this, add in valid name checking per string in a path, as transform names that contain spaces currently cause
geometry name matching for `materialassign`s to incorrectly match (this is a separate issue that can be raised be addressed.  e.g. "/ blah" will match all geometry with any "/" in it, which could be everything. )

3. Update default shaderball to remove a top level transform called "input.obj". No other changes were made to the file. The change makes both meshes reside at the root of the scene, which in turns matches the path specifiers used for the example materials file which use "look" material assignments (std surface brass and wood).