Support for new OSL closures in MaterialXGenOsl (#1039)
This change list adds a build option to select between supporting either the old legacy OSL closures or the new OSL closures in development. By default the old closures are selected for backwards compatibility.

To select the new closures one needs to disable the build option MATERIALX_OSL_LEGACY_CLOSURES and use a build of OSL that has the new closures supported in testrender. Also make sure to remove the ifdef MATERIALX_CLOSURE guards in stdosl.h to make the new closures visible when compiling the generated shaders.