Adds a Unified Noise node
Adds a unified noise node, for artists to have a single noise node they can create, which unifies common controls (such as frequency and offset).  A common complaint we hear from artists, is the MaterialX noises are so bare, they are missing the very basics of frequency/offset, and even some post-processing, which requires creating all sorts of nodes.

The fractal controls are unique to that particular noise type, so an artist tool would probably hide those parameters when the noise type is not set to fractal.

Jitter is treated as a common control; it behaves like jitter does on the standard worley noise, and for the non-Worley noises it scrambles the input.