Adds a Unified Noise node (#1087)
Adds a unified noise node, for artists to have a single noise node that unifies shared controls (such as frequency and offset). A common complaint we hear from artists, is the MaterialX noises are too bare, they are missing the very basics of frequency/offset, and even some post-processing, which requires creating all sorts of nodes. Rather than modify all of the existing noises, it seemed a slightly higher-level node to encompass shared features made more sense (and would be easier to adopt).

Jitter is treated as a common control; it behaves like jitter does on the standard worley noise, and for the non-Worley noises it scrambles the input. The fractal controls are unique to that particular noise type, so an artist tool would probably hide those parameters when the noise type is not set to fractal.