Fix "unbind geometry" MaterialXView OpenGL errors on MacOS (#1126)
In GlslProgram::unbindGeometry, glDisableVertexAttribArray is called. According to the specification (https://registry.khronos.org/OpenGL-Refpages/gl4/html/glEnableVertexAttribArray.xhtml), this requires a vertex array to be bound, which is not the case. On MacOS, the driver seems to be less lenient and returns an error which is then printed as "OpenGL error after program unbind geometry: 1282".

Since we destroy the vertex array a few lines later, there's no need to disable the attributes.

Removing the glDisableVertexAttribArray calls fixes the errors.