Fix installation issue with OSL legacy closures
When building with MATERIALX_OSL_LEGACY_CLOSURES=ON CMake would install
the non-legacy pbrlib_genosl_impl.mtlx first and then it would find
that it is already up-to-date and not install the correct legacy
on top of it. This commit fixes this by excluding both the legacy
and the non-legacy file from the glob install and then installing
the correct one explicitly.

Note that if the user changes the value of MATERIALX_OSL_LEGACY_CLOSURES
and then installs again the correct file will NOT be installed.