Add node for directionally varying emission. (#1262)
This change list adds a new node to pbrlib: generalized_schlick_edf.

This node can be used to create a directionally varying / Fresnel-like modulation to emission, using the generalized Schlick Fresnel curve. The main idea is to use this to approximate how emission is attenuated when placed under a coating, which is supported in shading models like Standard Surface.