Fix value types for swizzles in shader generation (#1335)
For the swizzling logic on inputs in the base shader generator, the returned type of the Value after swizzling can be incorrect as it looks up the type based on the Value's Type as mapped to a TypeDesc -- as opposed just the Type. The TypeDesc type can end up not being a valid Type which results in a default string Value being returned.

This occurs for the "channels" conversion for instance for "color3" for MSL and GLSL as the TypeDesc type is "vec3" instead of the correct type which is "color3".