Cleanup old unused code for conditional nodes (#1344)
This change list removes some old unused code paths related to conditional node handling in shader generation.

In the early days of ShaderGen we added a feature to optimize how code was generated for conditional nodes. In particular the goal was to identify sub-graphs that were only used by certain branches of conditionals, and restrict codegen of these nodes to be only emitted inside the scope of the corresponding branches. However the feature as implemented were never made 100% safe and has been disabled for many years.

Since the code is unused, unsafe, and since shader compilers will handle this later anyway on the generated shader, I think it's about time to clean this up.