Define default color transforms as graphs
This changelist updates the definitions of the default color transforms from raw shader code to MaterialX graphs, removing the need for explicit code for each transform to be provided in each supported language.

Further in the future, this approach could enable direct communication of color transform graphs from OpenColorIO to MaterialX, allowing both standard ACES transforms and custom transforms to be communicated in a unified way.