Remove unneeded clamp in UsdPreviewSurface
This changelist removes an unneeded clamp in the graph definition of UsdPreviewSurface, providing a small simplification to its generated shader code in all languages.

The formal specification for UsdPreviewSurface may be found at https://openusd.org/release/spec_usdpreviewsurface.html, and this clamp is not required to match its expected behavior for opacity.