Merge supplemental nodes into main specification (#1377)
Descriptions of all of the "Supplemental Nodes" are moved from the Supplemental Notes doc to the main Specification doc, as discussed in various forums. The previously-supplemental nodes now have a subtle "(NG)" identifier in the description, and the text has been updated to explain this difference (There are likely several original standard nodes that should get this "(NG)" identifier).
The README file was updated to reflect the above change, as well as a few previous Spec updates that weren't yet reflected in the README.
I also added link target elements for each individual Merge node so they can be linked to from the outside.