Generate Python wheels in CI (#1317)
This PR adds a new GitHub Actions workflow that generates Python wheels and an sdist that all can be uploaded to PyPI.

It generates wheels for Python 3.7, 3.8, 3.9 3.10, and 3.11 for Windows, macOS and Linux.

I used https://github.com/scikit-build/scikit-build-core as the build backend. It's a new tool that is quite simple and allows to build Python extensions (C, C++, etc) without setuptools. Since it's a build backend, everything can be built with pip install. or https://github.com/pypa/build (python -m build).