Refine dot node elision and remove unused code (#1522)
Dot nodes were elided in #1152 to prevent creating extra samplers in rasterized shadergen.

This preserves the change written by @kwokcb while keeping the distinction between a topological constant node and a non-topological dot node.

In scenarios where you want to feed a single color to multiple shader nodes in a complex material, using a constant node will force recompilation every time the color is tweaked, we want users to be able to use a dot node to feed the graph, which should allow tweaking the color without having to recompile the shader. Also allows re-using the shader code for material instances differing on this input color.