Fix issue #1464
This merges both `TexCoordNodeMsl` and `TexCoordNodeGlsl` into a single
`HwTexCoordNode`. The parts of the node that are configurable (e.g. the
input index) can be changed via re-implementation of the virtual member
function `getIndexInput()`.

The new mechanism works regardless of syntax, and uses the smallest
texcoord width, depending on what is used in the shader graph.