Add fileTextureVerticalFlip option for MDL generation (#1418)
When investigating the differences in the MaterialXTest I implemented the fileTextureVerticalFlip in the MDL shader gen.
In theory this option is not needed because the MaterialX and MDL define the image coordinate both in the lower left.
But it might be useful to align with existing but non standard conform renderers. It could for instance be used in the MaterialXTest to align with the flipped GLSL and OSL results.