Improve "Add Node" workflows in the graph editor
Previously, the tab menu was the only way to add new nodes to the graph editor.
Now users can select any pin in the graph, and drag a link off of that pin. If
this link does not connect to another pin, on mouse release, the user will be
prompted to add a new node. The list of nodes that users can add will be filtered
to only include nodes that can be connected to the pin from which this link originates.

This context-aware node filtering in the "Add Node" popup can be toggled on/off
with the "Smart Filter" check box.