cmake: factor out COMMON_LIBRARIES var from RenderMsl and RenderGlsl
The point of COMMON_LIBRARIES was apparently to allow all library
linking to be done with a single command; however, continuing this
paradigm becomes difficult as more complex conditions are built up -
for instance, in MaterialXRenderMsl/CMakeLists.txt, we were altering
COMMON_LIBRARIES if APPLE and NOT iOS... which, at best, made the name
COMMON_LIBRARIES unclear.

Refactored to simply make multiple calls to target_link_libraries, which
is the way most other CMakeLists.txt in this repo handle this.