Upgrade switch node to ten inputs in 1.39 (#1660)
Also added an upgrade function, which may need some closer attention. The upgrade function attempts to retain the behavior from 1.38 where any value of which that is above the accepted range that would previously been served the default (in1) value. A which value of 6 would have previously return in1, but now it would return one of the newly added inputs. To retain the previous renderer image, the document upgrader sets the value of which to the default of 0.