Redesign of TypeDesc class (#1711)
This change list implements a redesign of the TypeDesc class used for data type descriptions in MaterialX.

The main purpose of this change is to solve issue #1608. But the new design also gives performance improvements to the use of this class.

Since the TypeDesc class is used frequently many files are touched by this change and there are some API breakage. Function wrappers with [[deprecated]] are added to minimize API breakage, but it's not possible for everything that changed. For example, the TypeDesc instances are now stored and accessed by value rather then by pointer, requiring a change to the access syntax.