Initial commit legacy materials (#1398)
* Initial commit legacy materials

This commit includes:
 - all Legacy (Protein) classes
 - procedural nodes (checker, noise, turbulence 2d/3d)
 - utilities (stain, normals combine nodes)

For now all nodes are under "adsk_legacy" group, except the turbulence ones that might be ok to be added to the standard lib.

Becasue of a bug in the handling of namespaces in the Graph Editor all namespaces are set to "". Bug: https://github.com/AcademySoftwareFoundation/MaterialX/issues/1695

* Test files for legacy material

This includes a single MTLX file and a set of textures.
The MTLX file can be loaded in MaterialX Viewer and materials selected from the UI

* Adding values to test file

Adding values to all inputs to avoid warnings. This files generates no warning messages  when loaded in MaterialX viewer

Also fixing a small typo in the defs file

* USD test file

This is a USD test file that can be loaded in usdview to test the same materials that the legacy_materials_test.mtlx tests in MaterialX viewer.

Open the file in usdview, select the ShaderBall object and in the  Meta data panel you can select a variant for each material class.

The legacy materials node defs are also updated to remove a stub for namespace that was causing trouble in usdview (but was ok in MaterialX viewer).
A typo in the comments was also fixed.

* Stubs for OSL and MDL turbulence noise

Stubs for implementation of:
 - turbulence2d_float
 - turbulence3d_float
in OSL and MDL.
Currently returning a solid color. This is just to check the build warnings.

* Test for matching implementation count

Just a test to see if hardcoded expected skip count does influence overall implementation counters and failure to match.
Testing for glsl only.

* Revert "Test for matching implementation count"

This reverts commit 12e011f641750b2b6309defd60fbe1a95aa3e7be.

* Added test mtlx file for unused nodedefs

Tests were failing because the node turbulence2D was defined but never used in any TestSuite file.

Adding a other.mtlx as a container for these orphans nodes.
Note that for consistency I also added turbulence3d in there, even if not necessary as it's used in the main material test file.