Enable color and unit transforms on root nodes
This changelist adds support for color and unit transforms on the root nodes that are passed into shader generation, where previously they were only supported on non-root nodes in the material graph.

In MaterialX 1.38.9, a special-case pathway for root surface shader nodes was merged into the main logic, causing color transforms in some USD/MaterialX contexts to be omitted, and this changelist restores and generalizes the original approach to handle all root nodes.