Fix redundant values
Due to a bug in graph editor, a lot of inputs had connections to an interfacename and a value, when  only the interfacename was needed.

This is fixed in future graph editor builds, but probably only for new created inputs, so a clean slate is necessary for safety.

The graphs were working in Viewer, but causing issues in LookdevX. LookdevX now handles this case if we face it again in the future.