Add Python 3.12 to wheels generation (#1802)
The highest version of Python for generating wheels in the Github workflow is currently 3.11. However, Python 3.12 is now quite widespread (Ubuntu 24.04, Arch Linux, Windows etc.). On a platform with Python 3.12, pip install results in recompilation from source, which is quite tedious.

The proposal is therefore simply to extend the generation of wheels in the Github workflow to Python 3.12, to feed PyPI.