Initial support for Zeltner sheen
This is a first version of the GLSL implementation, which is currently inactive but can be enabled in place of the existing sheen model via the SHEEN_METHOD #define in mx_sheen_bsdf.glsl.

In contrast to the reference implementation (https://github.com/tizian/ltc-sheen), this version uses analytic fits instead of a table, to ease integration. I plan to improve the quality of the fits in the future (particularly the directional albedo), and also add a sampling function.

Remove debug code