Default MATERIALX_BUILD_TESTS to false
This changelist updates the default for the `MATERIALX_BUILD_TESTS` build option from `ON` to `OFF`, as this option generates a relatively heavyweight cache of the `libraries` and `resources` folders.

Often, developers want to build and test the MaterialX Viewer or MaterialX Graph Editor, and in these situations it's more natural for `MATERIALX_BUILD_TESTS` to be a feature that is opt-in rather than opt-out.