Fix OSL syntax for Convert nodes (#1863)
PR #1854, previously removed the C++ implementations for `convert` and replaced them with direct source code implementations.  The testing done wasn't correctly running `oslc` so a few syntax errors slipped through.

These are corrected in this PR.

Note:
`vector2`, `vector4` and `color4` are implemented as structs in OSL, and so do not have constructors in the typical sense.   The following is not legal OSL syntax.
```
vector4 a = vector4(0,0,0,0);
```
instead you must use a list initializer to initialize the struct.
```
vector4 a = {0,0,0,0};
```

Conversely, currently it is NOT legal to initialize a builtin type, for instance `color` or `vector` using a list initializer. 
The following is not legal OSL syntax.
```
color a = {0,0,0};
```
instead you must use the type constructor.
```
color a = color(0,0,0);
```

Also note, subscript access also is also not legal OSL syntax because these types are implemented as structs. For example to access the green component of a `color4`, you must use `myColor.rgb[1]`.