Remove unused thin_surface node (#1890)
This changelist removes the `thin_surface` node. The specification for thin-walled surfaces was changed in 1.39, making this node deprecated. 

The node was never implemented across multiple targets (MDL is the only exception), and we've never heard of anyone trying to use it. So no upgrade logic is added at this point. Should anyone need an upgrade path we can add that later.